kiwi_cook_server/pyproject.toml
[tool.poetry]
name = "kiwi_cook_server"
version = "0.1.0"
description = ""
authors = ["am9zZWY <46693545+am9zZWY@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aiohttp = "^3.11.0"
black = "^24.10.0"
cachetools = "^5.5.0"
fastapi = "^0.115.2"
httptools = "^0.6.4"
ingredient-parser-nlp = "^1.3.0"
openai = "^1.53.0"
pip-chill = "^1.0.3"
pydantic = "^2.9.2"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
recipe-scrapers = "^15.2.1"
starlette = "^0.39.2"
uvloop = "^0.21.0"
uvicorn = "^0.32.0"
pint = "^0.24.4"
python-multipart = "^0.0.17"
argon2-cffi = "^23.1.0"
pyjwt = "^2.9.0"
rapidfuzz = "^3.10.1"
pytest = "^8.3.3"
fastapi-limiter = "^0.1.6"
redis = "^5.2.0"
passlib = "^1.7.4"

[tool.poetry.group.dev.dependencies]
fastapi = "^0.115.2"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
